% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{copy_template}
\alias{copy_template}
\title{Copy template}
\usage{
copy_template(pth, template_dir, ignore = "R/templates.R")
}
\arguments{
\item{template_dir}{string. The name of the template dir to copy files from.}
}
\description{
Copies template files from www and the specified template directory
}
