% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{create_dashboard}
\alias{create_dashboard}
\title{Create dashboard}
\usage{
create_dashboard(
  pth,
  template_dir = "app_template",
  dirs = list.dirs(c("R", "www"))
)
}
\arguments{
\item{pth}{string. The path root directory of the application.}
}
\description{
Creates an application set-up that includes TPL styles in the css folder
}
