% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{create_report}
\alias{create_report}
\title{Create report}
\usage{
create_report(
  pth,
  template_dir = "report_template",
  dirs = list.dirs(c("R", "www"))
)
}
\arguments{
\item{pth}{string. The path root directory of the application.}
}
\description{
Creates a rmarkdown report set-up that includes TPL styles in the css folder
}
