% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer.R
\name{footer}
\alias{footer}
\title{TPL footer}
\usage{
footer(TPL_URL = "https://www.texaspolicylab.org/")
}
\arguments{
\item{version}{the version of the dashboard}
}
\description{
Creates the TPL footer to add to dashboards and reports
}
