% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{test_list_item_not_null}
\alias{test_list_item_not_null}
\title{Test list item is not null}
\usage{
test_list_item_not_null(l, item_name)
}
\arguments{
\item{l}{the list the check}

\item{item_name}{the item name in the list to check}
}
\description{
Test list item is not null
}
